+++ examples/commonjs/App.vue
@@ -1,10 +1,18 @@
 <template>
-  <div>
+  <div style="background: #000;padding:50px;">
     <star-rating
       v-model:rating="rating"
-      :increment="0.5"
       :round-start-rating="false"
+      :star-points="[23,2, 14,17, 0,19, 10,34, 7,50, 23,43, 38,50, 36,34, 46,19, 31,17]"
+      :border-width="3"
+      border-color="#fff"
+      :padding="5"
       :clearable="true"
+      inactive-color="#fff"
+      :apply-active-color-on-hover="true"
+      :animate="true"
+      glow-color="#fff"
+      :glow="5"
     />
     <a
       href="#"
@@ -22,7 +30,7 @@ export default {
     },
     data(){
     	return{
-    		rating: 3.5
+    		rating: 0
     	}
     }
 }